Implement ESC [ sequences in vt100 mode

+++ vram.h
@@ -7,5 +7,5 @@ void vram_prev(void);
 void vram_next(void);
 void vram_ptr(u8 x, u8 y, u8 nrows);
 
-void vram_write(u8 c, u8 att);
+void vram_write(u16 c, u8 att);
 #endif /* _VRAM_H_ */
